Fire event for packets received from remotes.
When a packet with a remote command is received, fire a Home Assistant
event that contains data about the unit, channel and command.  This is
done in addition to the existing logic for processing those commands
(e.g. to sync cover state to remote commands), so that the events can
be used separately for automations or other purposes if desired.